Live analysis updates

+++ BookApp/Server/Services/HighlightServerService.cs
@@ -4,11 +4,13 @@
     {
         private readonly IHighlightRepository _highlightRepository;
         private readonly IHighlightMapper _highlightMapperService;
+        private readonly IHubServerService _hubServerService;
 
-        public HighlightServerService(IHighlightRepository highlightRepository, IHighlightMapper highlightMapperService)
+        public HighlightServerService(IHighlightRepository highlightRepository, IHighlightMapper highlightMapperService, IHubServerService hubServerService)
         {
             _highlightRepository = highlightRepository;
             _highlightMapperService = highlightMapperService;
+            _hubServerService = hubServerService;
         }
 
         public async Task<ServiceResponse> AddHighlight(HighlightModel newHighlight)
@@ -16,27 +18,23 @@
             var mappedHighlight = await _highlightMapperService.MapToDbModel(newHighlight);
             var addedHighlight = await _highlightRepository.Create(mappedHighlight);
             var mappedHighlightModel = await _highlightMapperService.MapToClientModel(addedHighlight);
+
+            await _hubServerService.HighlightAdded(mappedHighlightModel);
             return ServiceResponse<HighlightModel>.Success(mappedHighlightModel, "Highlight created.");
         }
 
         public async Task<ServiceResponse> DeleteHighlight(int highlightId)
         {
             var highlightToDelete = await _highlightRepository.FindByConditionsFirstOrDefault(h => h.Id == highlightId);
-            await _highlightRepository.Delete(highlightToDelete);
-
-            return ServiceResponse.Success("Highlight deleted.");
-        }
-
-        public async Task<ServiceResponse> RetrieveHighlights(int bookAnalysisId)
-        {
-            var highlights = await _highlightRepository.FindByConditions(h => h.BookAnalysisId == bookAnalysisId);
-            List<HighlightModel> mappedHighlights = new();
-            foreach (Highlight highlight in highlights)
+            if (highlightToDelete is null)
             {
-                mappedHighlights.Add(await _highlightMapperService.MapToClientModel(highlight));
+                return ServiceResponse.Error("Highlight not found.");
             }
 
-            return ServiceResponse<List<HighlightModel>>.Success(mappedHighlights, "Highlights retrieved.");
+            await _highlightRepository.Delete(highlightToDelete);
+
+            await _hubServerService.HighlightRemoved(highlightToDelete.BookAnalysisId, highlightId);
+            return ServiceResponse.Success("Highlight deleted.");
         }
 
         public async Task<ServiceResponse> UpdateHighlight(HighlightModel updatedHighlight)
@@ -50,6 +48,7 @@
             _highlightMapperService.MapEditHighlight(highlightToUpdate, updatedHighlight);
             await _highlightRepository.Edit(highlightToUpdate);
 
+            await _hubServerService.HighlightUpdated(updatedHighlight);
             return ServiceResponse.Success("Highlight updated.");
         }
     }
